Update ParkingLotTest
You don't return anything in the tests(such as boolean), you either fail them or you don't, if it doesn't fail then it counts as passing the test. 
Also to check if it's full its best to add all the possible cars(200 as a place holder until we decide on a max amount of cars) and then check if the lot has 200 cars.