Update ParkingLot
Removed the redundant size to the ArrayList(200), an ArrayList will keep adding forever even with that there. To counteract that I changed the addCar method to stop adding when it reaches 200 cars in the ArrayList. Also the isFull method's if statement was a little odd(though still worked probably) so I made it simpler.